gardenapp_done

+++ week-05/day-02/03-GardenApp_Rework/flower.h
@@ -5,7 +5,7 @@
 
 class Flower : public Plant {
 public:
-    Flower(int currentWaterAmount, const std::string &color);
+    Flower(double currentWaterAmount, const std::string &color);
     void absorbWater(double getWaterAmount) override;
     bool needWater() override;
