+++ week-05/day-02/03-GardenApp_Rework/tree.cpp
@@ -1,6 +1,6 @@
 #include "tree.h"
 
-Tree::Tree(int currentWaterAmount, const std::string &color) : Plant(currentWaterAmount, color) {
+Tree::Tree(double currentWaterAmount, const std::string &color) : Plant(currentWaterAmount, color) {
     _name = "Tree";
 }
