+++ week-05/day-02/03-GardenApp_Rework/tree.h
@@ -5,7 +5,7 @@
 
 class Tree : public Plant {
 public:
-    Tree(int currentWaterAmount, const std::string &color);
+    Tree(double currentWaterAmount, const std::string &color);
     void absorbWater(double getWaterAmount) override;
     bool needWater() override;
